Create User model and run it's associated migration
The `Landlord` and `Tenant` models do not sufficiently represent
how the SunCity system will be used:
Users of the system will be able to view and rent accommodations
on the system as well as rent their spaces out to other users,
thus acting as both tenants and landlords at the same time.

Having two seperate Rails models for Tenant and Landlord means
that data for users who are both landlords and tenants will be
stored in two different database tables, and this defeats the
DRY principle, as well as creating future complications as to
where data should be retrieved.

In view of this and other considerations, it is best that users
be represented by a single `Users` model which will store all info
for a user whether it is Tenant info or Landlord info.

The Users model is created using:
 `bin/rails generate model User first_name last_name \
  email phone_number`
after which `bin/rails db:migrate` is run