Remove everything related to Tenants and Landlords
As stated in a previous commit, the Tenant and Landlord model
have been replaced with the User model. This commit therefore
gets rid of the Tenant and Landlord models and their associated
references (in other models), controllers, views and resources.

In addition, 2 migrations are created for dropping the tenants and
landlord tables respectively